Task_A/requirements.txt
# Core PyTorch with CUDA support
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Model architectures and utilities
timm>=0.9.0
transformers>=4.30.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=9.0.0
albumentations>=1.3.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.9.0

# Data Processing
pandas>=1.5.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress and Logging
tqdm>=4.64.0
wandb>=0.15.0
tensorboard>=2.13.0

# Image Processing
imageio>=2.22.0
scikit-image>=0.19.0

# Optimization and Acceleration
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Mixed Precision Training
apex>=0.1

# Model Compression
torch-pruning>=1.2.0

# Metrics and Evaluation
torchmetrics>=0.11.0

# Utilities
psutil>=5.9.0
rich>=13.0.0
click>=8.1.0
pyyaml>=6.0

# Jupyter (optional for development)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Memory optimization
py3nvml>=0.2.7
